requirements.txt
discord.py
aiohttp
pystray
pillow
